requirements.txt
flask
flask_sqlalchemy
flask_migrate
flask_mail
python-dotenv
pymysql
mysql-connector-python
apscheduler
reportlab
requests
sentence-transformers
scikit-learn
numpy
google-cloud-aiplatform>=1.51.0
google-api-python-client
google-auth
google-auth-oauthlib
flask-limiter
transformers
torch
sentencepiece
safetensors
